LPS-14310 expiration in millisec, need long instead of int
git-svn-id: svn://svn.liferay.com/repos/public/plugins/trunk@70393 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portlets/opensocial-portlet/docroot/WEB-INF/src/com/liferay/opensocial/service/impl/OAuthTokenLocalServiceImpl.java
@@ -33,7 +33,7 @@ public class OAuthTokenLocalServiceImpl
 	public OAuthToken addOAuthToken(
 			long userId, long gadgetId, String serviceName, long moduleId,
 			String accessToken, String tokenName, String tokenSecret,
-			String sessionHandle, int expiration)
+			String sessionHandle, long expiration)
 		throws PortalException, SystemException {
 
 		User user = UserLocalServiceUtil.getUser(userId);
